+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This release uses a new Raft consensus version, with the move to Hashicorp Raft 
 - [PR #598](https://github.com/rqlite/rqlite/pull/598): Ensure backup is correctly closed.
 - [PR #600](https://github.com/rqlite/rqlite/pull/600): Move to Hashicorp Raft v1.
 - [PR #601](https://github.com/rqlite/rqlite/pull/601): By default use Raft network address as node ID.
+- [PR #602](https://github.com/rqlite/rqlite/pull/602/files): Add method to Store that returns leader ID.
 
 ## 4.6.0 (November 29th 2019)
 _This release adds significant new functionality to the command-line tool, including much more control over backup and restore of the database. [Visit the Releases page](https://github.com/rqlite/rqlite/releases/tag/v4.6.0) to download this release._